Refactor packge specs and resolution
We now have an actually somewhat reasonable package version resolution
method.

I also split out package requests and package specs, so that the type
system prevents us from using a package request with a version that
hasn't been resolved yet in a place where we expect it to be resolved.

In the future I might also split the package spec into one for known and
one for installed packages.