Decouple installed packages from workspaces on disk
With this change, we store installed packages separately under
~/.local/matcha/packages and link them into ~/.local/matcha/workspaces.

In the current version, we also never actually delete any installed
packages, even when they're not in use, which we will want to do
eventually.

This design is closer to what Nix does, and also allows us to reuse
already installed packages across multiple workspaces in the future.