Add GC for installed but unused packages
Doing this as an explicit step for now, because I don't want to do the
thing that brew does where I tell the program to do one thing and it
does a bunch of other things that I didn't ask it to do.

Also this way the user gets to decide if they want to just keep old
versions around for a while to e.g. rollback easier.