Checkpoint: Use package requests for updates as well
This gradually moves more code from the package module to the
dependencie module. The eventual plan is to merge the two, get rid of
PackageRequest, and have DependecyRequest take its place.

I'm also gradually removing the progress bar code, because I realised
progress bars are for slow programs, and we're writing a fast program.