Fix a bug that created duplicate workspace packages
If one had a version of a package installed in the a workspace, and then
tried to install the package again without specifying a version, we
would install the newer version of the package in the workspace, but not
through the upgrade path. This would prevent the old version from being
deregistered.

It would also cause all subsequent package resolutions to fail, because
two conflicting version requests would be present in the workspace.